changed .gitignore to make git not ignore .iml file so that other can pull the project correctly. But not tested if the changed worked.
Downloaded two icon in the drawable. Their source is recorded in the download file source.txt.

Changed layout of vmselection so the floating action bar was removed.

Changed layout of list_item_machines in order to separate the text and icon. But didn't change how the layout was inflated yet. It appears require to changed the adapter.