Just added some comments and thoughts about receiving data.
Also, another thing I have considered is sync adapter. For now, I think sync adapter is not necessary.
No matter what, GCM will need to notify the app that the data in database is changed.
So I didn't see the point why not just send the data directly with the first notice instead of notifying the sync adapter.